// types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export interface Database {
  public: {
    Tables: {
      Todos: {
        Row: {
          created_at: string | null
          id: number
          is_done: boolean | null
          text: string | null
        }
        Insert: {
          created_at?: string | null
          id?: number
          is_done?: boolean | null
          text?: string | null
        }
        Update: {
          created_at?: string | null
          id?: number
          is_done?: boolean | null
          text?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}